refactor(tui): convert ui.rs to ui/ module directory with view submodules
Split monolithic ui.rs (2100 lines) into focused submodules:
- ui/mod.rs: render() entry point and dispatcher (~130 lines)
- ui/main_view.rs: main kanban view renderers
- ui/settings_view.rs: settings view renderer
- ui/card_detail.rs: card detail view and relationship boxes
- ui/board_detail.rs: board detail view
- ui/sprint_detail.rs: sprint detail view renderers
- ui/dialogs.rs: thin dialog/popup wrapper functions
