+++ crates/kanban-tui/src/components/mod.rs
@@ -1,23 +1,33 @@
 pub mod banner;
 pub mod card_detail_sections;
+pub mod conflict_popup;
 pub mod card_list_item;
 pub mod detail_view;
+pub mod filter_popup;
+pub mod footer;
 pub mod generic_list;
+pub mod help_popup;
 pub mod list;
 pub mod panel;
 pub mod popup;
+pub mod relationship_popup;
 pub mod relationship_section;
 pub mod selection_dialog;
 pub mod selection_list;
 
 pub use banner::*;
 pub use card_detail_sections::*;
+pub use conflict_popup::*;
 pub use card_list_item::*;
 pub use detail_view::*;
+pub use filter_popup::*;
+pub use footer::*;
 pub use generic_list::*;
+pub use help_popup::*;
 pub use list::*;
 pub use panel::*;
 pub use popup::*;
+pub use relationship_popup::*;
 pub use relationship_section::*;
 pub use selection_dialog::*;
 pub use selection_list::*;
