--- crates/kanban-tui/src/ui.rs
@@ -1,2100 +1,0 @@
-use crate::app::{App, AppMode, BoardFocus, CardFocus, DialogMode, Focus};
-use crate::components::*;
-use crate::theme::*;
-use crate::view_strategy::UnifiedViewStrategy;
-use kanban_domain::{dependencies::CardGraphExt, Sprint, SprintStatus};
-use ratatui::{
-    layout::{Constraint, Direction, Layout, Rect},
-    style::{Color, Modifier, Style},
-    text::{Line, Span},
-    widgets::{Block, Borders, Clear, ListItem, Paragraph},
-    Frame,
-};
-use uuid::Uuid;
-
-// Card detail view layout constants
-const RELATIONSHIP_BOX_HEIGHT: u16 = 7;
-const RELATIONSHIP_VIEWPORT_BORDER_HEIGHT: usize = 2;
-
-fn render_banner(app: &App, frame: &mut Frame, area: Rect) {
-    if let Some(ref banner) = app.ui_state.banner {
-        banner.render(frame, area);
-    }
-}
-
-pub fn render(app: &mut App, frame: &mut Frame) {
-    // Check if we're in Help mode and render underlying view
-    let is_help_mode = matches!(app.mode, AppMode::Help(_));
-
-    if !is_help_mode {
-        let chunks = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([Constraint::Min(0), Constraint::Length(3)])
-            .split(frame.area());
-
-        // Phase 1: Render base view (from stack if in dialog mode)
-        let base_mode = app.get_base_mode();
-        match base_mode {
-            AppMode::CardDetail => render_card_detail_view(app, frame, chunks[0]),
-            AppMode::BoardDetail => render_board_detail_view(app, frame, chunks[0]),
-            AppMode::SprintDetail => render_sprint_detail_view(app, frame, chunks[0]),
-            AppMode::Settings => render_settings_view(app, frame, chunks[0]),
-            _ => render_main(app, frame, chunks[0]),
-        }
-        render_footer(app, frame, chunks[1]);
-
-        // Phase 2: Render dialog overlay if active
-        if let AppMode::Dialog(ref dialog) = app.mode {
-            match dialog {
-                DialogMode::CreateBoard => render_create_board_popup(app, frame),
-                DialogMode::CreateCard => render_create_card_popup(app, frame),
-                DialogMode::CreateSprint => render_create_sprint_popup(app, frame),
-                DialogMode::RenameBoard => render_rename_board_popup(app, frame),
-                DialogMode::ExportBoard => render_export_board_popup(app, frame),
-                DialogMode::ExportAll => render_export_all_popup(app, frame),
-                DialogMode::ImportBoard => render_import_board_popup(app, frame),
-                DialogMode::SetCardPoints => render_set_card_points_popup(app, frame),
-                DialogMode::SetCardPriority => render_set_card_priority_popup(app, frame),
-                DialogMode::SetMultipleCardsPriority => {
-                    render_set_multiple_cards_priority_popup(app, frame)
-                }
-                DialogMode::SetBranchPrefix => render_set_branch_prefix_popup(app, frame),
-                DialogMode::SetSprintPrefix => render_set_sprint_prefix_popup(app, frame),
-                DialogMode::SetSprintCardPrefix => render_set_sprint_card_prefix_popup(app, frame),
-                DialogMode::OrderCards => render_order_cards_popup(app, frame),
-                DialogMode::CreateColumn => render_create_column_popup(app, frame),
-                DialogMode::RenameColumn => render_rename_column_popup(app, frame),
-                DialogMode::DeleteColumnConfirm => render_delete_column_confirm_popup(app, frame),
-                DialogMode::SelectTaskListView => render_select_task_list_view_popup(app, frame),
-                DialogMode::FilterOptions => render_filter_options_popup(app, frame),
-                DialogMode::AssignCardToSprint => render_assign_sprint_popup(app, frame),
-                DialogMode::AssignMultipleCardsToSprint => {
-                    render_assign_multiple_cards_popup(app, frame)
-                }
-                DialogMode::ConflictResolution => render_conflict_resolution_popup(app, frame),
-                DialogMode::ExternalChangeDetected => {
-                    render_external_change_detected_popup(app, frame)
-                }
-                DialogMode::ConfirmSprintPrefixCollision => {}
-                DialogMode::ManageParents => render_manage_parents_popup(app, frame),
-                DialogMode::ManageChildren => render_manage_children_popup(app, frame),
-                DialogMode::CarryOverSprint => render_carry_over_sprint_popup(app, frame),
-                DialogMode::ExportBoards => render_export_boards_popup(app, frame),
-            }
-        }
-    } else {
-        // Help mode: render base view without footer, then help popup
-        let base_mode = app.get_base_mode();
-        match base_mode {
-            AppMode::CardDetail => render_card_detail_view(app, frame, frame.area()),
-            AppMode::BoardDetail => render_board_detail_view(app, frame, frame.area()),
-            AppMode::SprintDetail => render_sprint_detail_view(app, frame, frame.area()),
-            AppMode::Settings => render_settings_view(app, frame, frame.area()),
-            _ => render_main(app, frame, frame.area()),
-        }
-        app.view.last_frame_area = frame.area();
-        render_help_popup(app, frame);
-    }
-
-    // Render banner on top if present
-    if app.ui_state.banner.is_some() {
-        let banner_area = Rect {
-            x: 0,
-            y: 0,
-            width: frame.area().width,
-            height: 3,
-        };
-        render_banner(app, frame, banner_area);
-    }
-}
-
-pub fn render_settings_view(app: &App, frame: &mut Frame, area: Rect) {
-    use crate::app::SettingsFocus;
-    use crate::components::detail_view::{
-        metadata_line_selectable, metadata_line_styled, FieldSectionConfig,
-    };
-    use crate::theme::colors::SELECTED_BG;
-
-    let columns = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
-        .split(area);
-
-    let left_sections = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Percentage(80), Constraint::Percentage(20)])
-        .split(columns[0]);
-
-    let config_focused = app.focus.settings_focus == SettingsFocus::Configuration;
-    let config_file_focused = app.focus.settings_focus == SettingsFocus::ConfigFile;
-    let storage_focused = app.focus.settings_focus == SettingsFocus::Storage;
-
-    let config_location = kanban_service::config::effective_configuration_location(&app.app_config);
-
-    // Left top: Configuration
-    let config_section = FieldSectionConfig::new(" Configuration ")
-        .with_focus_indicator(" Configuration [1] ")
-        .focused(config_focused);
-    let config_block = config_section.block();
-    let is_config_selected =
-        |i: usize| config_focused && app.selection.settings_config.is_selected(i);
-    let mut config_lines = vec![
-        metadata_line_selectable(
-            "Configuration Format",
-            app.app_config.effective_configuration_format(),
-            is_config_selected(0),
-        ),
-        metadata_line_selectable(
-            "Configuration Location",
-            &config_location,
-            is_config_selected(1),
-        ),
-        metadata_line_selectable(
-            "Default Card Prefix",
-            app.app_config.effective_default_card_prefix(),
-            is_config_selected(2),
-        ),
-        metadata_line_selectable(
-            "Default Sprint Prefix",
-            app.app_config.effective_default_sprint_prefix(),
-            is_config_selected(3),
-        ),
-        metadata_line_selectable(
-            "Editing Format",
-            app.app_config.effective_editing_format(),
-            is_config_selected(4),
-        ),
-    ];
-    if app.has_data_file {
-        if app.cli_file_override {
-            config_lines.push(metadata_line_styled(
-                "Storage Backend",
-                &app.config_storage_backend,
-                Style::default().fg(Color::DarkGray),
-            ));
-            config_lines.push(metadata_line_styled(
-                "Storage Location",
-                &app.config_storage_location,
-                Style::default().fg(Color::DarkGray),
-            ));
-            config_lines.push(metadata_line_selectable(
-                "Active Storage Backend",
-                app.app_config.effective_storage_backend(),
-                is_config_selected(7),
-            ));
-            config_lines.push(metadata_line_selectable(
-                "Active Storage Location",
-                app.app_config.effective_storage_location(),
-                is_config_selected(8),
-            ));
-        } else {
-            config_lines.push(metadata_line_selectable(
-                "Storage Backend",
-                app.app_config.effective_storage_backend(),
-                is_config_selected(5),
-            ));
-            config_lines.push(metadata_line_selectable(
-                "Storage Location",
-                app.app_config.effective_storage_location(),
-                is_config_selected(6),
-            ));
-        }
-    }
-    let config_paragraph = Paragraph::new(config_lines).block(config_block);
-    frame.render_widget(config_paragraph, left_sections[0]);
-
-    // Left bottom: Config File
-    let config_file_section = FieldSectionConfig::new(" Config File ")
-        .with_focus_indicator(" Config File [2] ")
-        .focused(config_file_focused);
-    let config_file_block = config_file_section.block();
-    let is_cf_selected =
-        |i: usize| config_file_focused && app.selection.settings_config_file.is_selected(i);
-    let config_path_display = if config_location.is_empty() {
-        "(unknown)".to_string()
-    } else {
-        config_location.clone()
-    };
-    let config_exists =
-        !config_location.is_empty() && std::path::Path::new(&config_location).exists();
-    let status = if config_exists { "Loaded" } else { "Not found" };
-    let config_format = app.app_config.effective_configuration_format();
-    let config_file_lines = vec![
-        metadata_line_selectable("Path", &config_path_display, is_cf_selected(0)),
-        metadata_line_selectable("Status", status, is_cf_selected(1)),
-        metadata_line_selectable("Configuration Format", config_format, is_cf_selected(2)),
-    ];
-    let config_file_paragraph = Paragraph::new(config_file_lines).block(config_file_block);
-    frame.render_widget(config_file_paragraph, left_sections[1]);
-
-    // Right: Storage
-    let storage_section = FieldSectionConfig::new(" Storage ")
-        .with_focus_indicator(" Storage [3] ")
-        .focused(storage_focused);
-    let storage_block = storage_section.block();
-    let is_storage_selected =
-        |i: usize| storage_focused && app.selection.settings_storage.is_selected(i);
-    let file_path = app.persistence.save_file.as_deref().unwrap_or("(none)");
-    let backend = if app.persistence.save_file.is_some() {
-        app.app_config.effective_storage_backend()
-    } else {
-        "(none)"
-    };
-    let instance_id = app.ctx.state_manager.instance_id().to_string();
-    let export_selected = is_storage_selected(3);
-    let export_checkbox_style = if export_selected {
-        Style::default().fg(Color::Yellow).bg(SELECTED_BG)
-    } else {
-        Style::default().fg(Color::Yellow)
-    };
-    let export_text_style = if export_selected {
-        Style::default().fg(Color::White).bg(SELECTED_BG)
-    } else {
-        Style::default().fg(Color::White)
-    };
-    let storage_lines = vec![
-        metadata_line_selectable("File", file_path, is_storage_selected(0)),
-        metadata_line_selectable("Backend", backend, is_storage_selected(1)),
-        metadata_line_selectable("Instance ID", &instance_id, is_storage_selected(2)),
-        Line::from(""),
-        Line::from(vec![
-            Span::styled("  [x] ", export_checkbox_style),
-            Span::styled("Export Boards", export_text_style),
-        ]),
-    ];
-    let storage_paragraph = Paragraph::new(storage_lines).block(storage_block);
-    frame.render_widget(storage_paragraph, columns[1]);
-}
-
-fn render_export_boards_popup(app: &App, frame: &mut Frame) {
-    use crate::app::{ExportFormat, ExportStep};
-
-    let Some(ref dialog) = app.export_dialog else {
-        return;
-    };
-
-    match dialog.step {
-        ExportStep::SelectBoards => {
-            let inner = render_popup_with_block(frame, "Select Boards to Export", 60, 60);
-
-            let chunks = Layout::default()
-                .direction(Direction::Vertical)
-                .margin(1)
-                .constraints([Constraint::Min(0), Constraint::Length(1)])
-                .split(inner);
-
-            let items: Vec<Line> = app
-                .ctx
-                .boards
-                .iter()
-                .enumerate()
-                .map(|(i, board)| {
-                    let checkbox = if dialog.board_selections.get(i).copied().unwrap_or(false) {
-                        "[x] "
-                    } else {
-                        "[ ] "
-                    };
-                    let style = if i == dialog.cursor {
-                        Style::default()
-                            .fg(Color::Yellow)
-                            .add_modifier(Modifier::BOLD)
-                    } else {
-                        Style::default().fg(Color::White)
-                    };
-                    Line::from(Span::styled(format!("{}{}", checkbox, board.name), style))
-                })
-                .collect();
-
-            let list = Paragraph::new(items);
-            frame.render_widget(list, chunks[0]);
-
-            let hint = Paragraph::new(Line::from(vec![Span::styled(
-                "Space: toggle | a: all | Enter: next | Esc: cancel",
-                Style::default().fg(Color::DarkGray),
-            )]));
-            frame.render_widget(hint, chunks[1]);
-        }
-        ExportStep::ExportOptions => {
-            let inner = render_popup_with_block(frame, "Export Options", 60, 30);
-
-            let chunks = Layout::default()
-                .direction(Direction::Vertical)
-                .margin(1)
-                .constraints([
-                    Constraint::Length(1),
-                    Constraint::Length(1),
-                    Constraint::Length(2),
-                    Constraint::Length(1),
-                    Constraint::Min(0),
-                ])
-                .split(inner);
-
-            let filename_label = Paragraph::new(Line::from(vec![
-                Span::styled("Filename: ", Style::default().fg(Color::Cyan)),
-                Span::styled(&dialog.filename, Style::default().fg(Color::White)),
-                Span::styled("_", Style::default().fg(Color::Yellow)),
-            ]));
-            frame.render_widget(filename_label, chunks[0]);
-
-            frame.render_widget(Paragraph::new(""), chunks[1]);
-
-            let json_style = if dialog.format == ExportFormat::Json {
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD)
-            } else {
-                Style::default().fg(Color::White)
-            };
-            let sqlite_style = if dialog.format == ExportFormat::Sqlite {
-                Style::default()
-                    .fg(Color::Yellow)
-                    .add_modifier(Modifier::BOLD)
-            } else {
-                Style::default().fg(Color::White)
-            };
-            let json_radio = if dialog.format == ExportFormat::Json {
-                "(*)"
-            } else {
-                "( )"
-            };
-            let sqlite_radio = if dialog.format == ExportFormat::Sqlite {
-                "(*)"
-            } else {
-                "( )"
-            };
-
-            let format_line = Paragraph::new(Line::from(vec![
-                Span::styled("Format: ", Style::default().fg(Color::Cyan)),
-                Span::styled(format!("{} JSON  ", json_radio), json_style),
-                Span::styled(format!("{} SQLite", sqlite_radio), sqlite_style),
-            ]));
-            frame.render_widget(format_line, chunks[2]);
-
-            let hint = Paragraph::new(Line::from(vec![Span::styled(
-                "Tab: format | Enter: export | Esc: back",
-                Style::default().fg(Color::DarkGray),
-            )]));
-            frame.render_widget(hint, chunks[3]);
-        }
-    }
-}
-
-fn render_main(app: &mut App, frame: &mut Frame, area: Rect) {
-    let is_kanban_view = if let Some(idx) = app.selection.active_board_index {
-        if let Some(board) = app.ctx.boards.get(idx) {
-            board.task_list_view == kanban_domain::TaskListView::ColumnView
-        } else {
-            false
-        }
-    } else {
-        false
-    };
-
-    if is_kanban_view {
-        app.view.viewport_height = area.height.saturating_sub(2) as usize;
-        render_tasks_panel(app, frame, area);
-    } else {
-        let chunks = Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([Constraint::Percentage(30), Constraint::Percentage(70)])
-            .split(area);
-
-        app.view.viewport_height = chunks[1].height.saturating_sub(2) as usize;
-        render_projects_panel(app, frame, chunks[0]);
-        render_tasks_panel(app, frame, chunks[1]);
-    }
-}
-
-fn render_projects_panel(app: &App, frame: &mut Frame, area: Rect) {
-    let mut lines = vec![];
-
-    if app.ctx.boards.is_empty() {
-        lines.push(Line::from(Span::styled(
-            "No projects yet. Press 'n' to create one!",
-            label_text(),
-        )));
-    } else {
-        for (idx, board) in app.ctx.boards.iter().enumerate() {
-            let config = ListItemConfig::new()
-                .selected(app.selection.board.get() == Some(idx))
-                .focused(app.focus.active == Focus::Boards)
-                .active(app.selection.active_board_index == Some(idx));
-
-            lines.push(styled_list_item(&board.name, &config));
-        }
-    }
-
-    let panel_config = PanelConfig::new("Projects")
-        .with_focus_indicator("Projects [1]")
-        .focused(app.focus.active == Focus::Boards);
-
-    let content = Paragraph::new(lines);
-    render_panel(frame, area, &panel_config, content);
-}
-
-pub fn build_filter_title_suffix(app: &App) -> Option<String> {
-    let mut filters = vec![];
-
-    if app.filter.hide_assigned_cards {
-        filters.push("Unassigned Cards".to_string());
-    }
-
-    if !app.filter.active_sprint_filters.is_empty() {
-        if let Some(board_idx) = app
-            .selection
-            .active_board_index
-            .or(app.selection.board.get())
-        {
-            if let Some(board) = app.ctx.boards.get(board_idx) {
-                let mut sprint_names: Vec<String> = app
-                    .ctx
-                    .sprints
-                    .iter()
-                    .filter(|s| app.filter.active_sprint_filters.contains(&s.id))
-                    .map(|s| s.formatted_name(board, "sprint"))
-                    .collect();
-                sprint_names.sort();
-                filters.extend(sprint_names);
-            }
-        }
-    }
-
-    if filters.is_empty() {
-        None
-    } else {
-        Some(format!(" - {}", filters.join(" + ")))
-    }
-}
-
-pub fn build_tasks_panel_title(app: &App, with_filter_suffix: bool) -> String {
-    let mut title = if app.mode == AppMode::ArchivedCardsView {
-        "Archive".to_string()
-    } else if app.focus.active == Focus::Cards {
-        "Tasks [2]".to_string()
-    } else {
-        "Tasks".to_string()
-    };
-
-    if with_filter_suffix && app.mode != AppMode::ArchivedCardsView {
-        if let Some(suffix) = build_filter_title_suffix(app) {
-            title.push_str(&suffix);
-        }
-    }
-
-    title
-}
-
-fn render_tasks_panel(app: &App, frame: &mut Frame, area: Rect) {
-    render_tasks(app, frame, area);
-}
-
-fn render_tasks(app: &App, frame: &mut Frame, area: Rect) {
-    if let Some(unified_strategy) = app
-        .view
-        .strategy
-        .as_any()
-        .downcast_ref::<UnifiedViewStrategy>()
-    {
-        unified_strategy
-            .get_render_strategy()
-            .render(app, frame, area);
-    }
-}
-
-fn render_sprint_detail_view(app: &App, frame: &mut Frame, area: Rect) {
-    if let Some(sprint_idx) = app.selection.active_sprint_index {
-        if let Some(sprint) = app.ctx.sprints.get(sprint_idx) {
-            if let Some(board_idx) = app.selection.active_board_index {
-                if let Some(board) = app.ctx.boards.get(board_idx) {
-                    let is_completed = sprint.status == SprintStatus::Completed;
-
-                    if is_completed {
-                        render_sprint_detail_with_tasks(app, frame, area, sprint, board);
-                    } else {
-                        render_sprint_detail_metadata(app, frame, area, sprint, board);
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn render_sprint_detail_metadata(
-    app: &App,
-    frame: &mut Frame,
-    area: Rect,
-    sprint: &Sprint,
-    board: &kanban_domain::Board,
-) {
-    let sprint_name = sprint.formatted_name(board, "sprint");
-
-    let mut lines = vec![
-        metadata_line_styled("Sprint", sprint_name, bold_highlight()),
-        Line::from(""),
-        metadata_line_styled(
-            "Status",
-            format!("{:?}", sprint.status),
-            sprint_status_style(sprint.status),
-        ),
-        metadata_line("Sprint Number", sprint.sprint_number.to_string()),
-    ];
-
-    if let Some(name) = sprint.get_name(board) {
-        lines.push(metadata_line("Name", name));
-    }
-
-    if let Some(start) = sprint.start_date {
-        lines.push(metadata_line(
-            "Start Date",
-            start.format("%Y-%m-%d %H:%M UTC").to_string(),
-        ));
-    }
-
-    if let Some(end) = sprint.end_date {
-        let end_style = if sprint.is_ended() {
-            Style::default().fg(Color::Red)
-        } else {
-            normal_text()
-        };
-        lines.push(metadata_line_styled(
-            "End Date",
-            end.format("%Y-%m-%d %H:%M UTC").to_string(),
-            end_style,
-        ));
-    }
-
-    lines.push(Line::from(""));
-
-    let card_count = app
-        .ctx
-        .cards
-        .iter()
-        .filter(|c| c.sprint_id == Some(sprint.id))
-        .count();
-    lines.push(metadata_line_styled(
-        "Cards Assigned",
-        card_count.to_string(),
-        Style::default()
-            .fg(Color::Cyan)
-            .add_modifier(Modifier::BOLD),
-    ));
-
-    if board.active_sprint_id == Some(sprint.id) {
-        lines.push(metadata_line_styled("Active Sprint", "Yes", active_item()));
-    }
-
-    lines.push(Line::from(""));
-
-    if let Some(prefix) = &sprint.prefix {
-        lines.push(metadata_line_styled("Sprint Prefix", prefix, active_item()));
-    }
-
-    if let Some(prefix) = &sprint.card_prefix {
-        lines.push(metadata_line_styled(
-            "Card Prefix Override",
-            prefix,
-            active_item(),
-        ));
-    }
-
-    if sprint.prefix.is_some() || sprint.card_prefix.is_some() {
-        lines.push(Line::from(""));
-    }
-
-    lines.push(metadata_line_styled(
-        "Created",
-        sprint.created_at.format("%Y-%m-%d %H:%M UTC").to_string(),
-        label_text(),
-    ));
-    lines.push(metadata_line_styled(
-        "Updated",
-        sprint.updated_at.format("%Y-%m-%d %H:%M UTC").to_string(),
-        label_text(),
-    ));
-
-    let content = Paragraph::new(lines).block(
-        Block::default()
-            .borders(Borders::ALL)
-            .border_style(focused_border())
-            .title("Sprint Details"),
-    );
-    frame.render_widget(content, area);
-}
-
-fn render_sprint_detail_with_tasks(
-    app: &App,
-    frame: &mut Frame,
-    area: Rect,
-    sprint: &Sprint,
-    board: &kanban_domain::Board,
-) {
-    let chunks = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
-        .split(area);
-
-    render_sprint_task_panel_with_selection(
-        app,
-        frame,
-        chunks[0],
-        sprint,
-        board,
-        &app.sprint_view.uncompleted_cards,
-        "Uncompleted",
-        app.sprint_view.panel == crate::app::SprintTaskPanel::Uncompleted,
-    );
-
-    render_sprint_task_panel_with_selection(
-        app,
-        frame,
-        chunks[1],
-        sprint,
-        board,
-        &app.sprint_view.completed_cards,
-        "Completed",
-        app.sprint_view.panel == crate::app::SprintTaskPanel::Completed,
-    );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn render_sprint_task_panel_with_selection(
-    app: &App,
-    frame: &mut Frame,
-    area: Rect,
-    _sprint: &Sprint,
-    board: &kanban_domain::Board,
-    task_list: &crate::card_list::CardList,
-    title_suffix: &str,
-    is_focused: bool,
-) {
-    let mut lines = vec![];
-    let selected_idx = task_list.get_selected_index();
-
-    if task_list.is_empty() {
-        lines.push(Line::from(Span::styled("  (no tasks)", label_text())));
-    } else {
-        let viewport_height = area.height.saturating_sub(2) as usize;
-        let render_info = task_list.get_render_info(viewport_height);
-
-        lines.extend(crate::scroll_indicators::render_above_indicator(
-            render_info.show_above_indicator,
-            render_info.cards_above_count,
-            "Task",
-        ));
-
-        for card_idx in &render_info.visible_card_indices {
-            if let Some(card_id) = task_list.cards.get(*card_idx) {
-                if let Some(card) = app.ctx.cards.iter().find(|c| c.id == *card_id) {
-                    let is_selected = selected_idx == Some(*card_idx) && is_focused;
-                    let animation_type = app
-                        .animation
-                        .animating
-                        .get(&card.id)
-                        .map(|a| a.animation_type);
-                    let line = render_card_list_item(CardListItemConfig {
-                        card,
-                        board,
-                        sprints: &app.ctx.sprints,
-                        is_selected,
-                        is_focused,
-                        is_multi_selected: false,
-                        show_sprint_name: false,
-                        animation_type,
-                        search_query: None,
-                    });
-                    lines.push(line);
-                }
-            }
-        }
-
-        lines.extend(crate::scroll_indicators::render_below_indicator(
-            render_info.show_below_indicator,
-            render_info.cards_below_count,
-            "Task",
-        ));
-    }
-
-    // Calculate points from cards in this panel
-    let cards: Vec<&kanban_domain::Card> = task_list
-        .cards
-        .iter()
-        .filter_map(|card_id| app.ctx.cards.iter().find(|c| c.id == *card_id))
-        .collect();
-    let points = kanban_domain::calculate_points(&cards);
-
-    lines.push(Line::from(Span::styled(
-        format!("Points: {}", points),
-        Style::default()
-            .fg(Color::Cyan)
-            .add_modifier(Modifier::BOLD),
-    )));
-
-    let border_style = if is_focused {
-        focused_border()
-    } else {
-        Style::default().fg(Color::DarkGray)
-    };
-
-    let content = Paragraph::new(lines).block(
-        Block::default()
-            .borders(Borders::ALL)
-            .border_style(border_style)
-            .title(format!("{} ({})", title_suffix, task_list.len())),
-    );
-    frame.render_widget(content, area);
-}
-
-fn render_footer(app: &App, frame: &mut Frame, area: Rect) {
-    let _is_kanban_view = if let Some(board_idx) = app
-        .selection
-        .active_board_index
-        .or(app.selection.board.get())
-    {
-        if let Some(board) = app.ctx.boards.get(board_idx) {
-            board.task_list_view == kanban_domain::TaskListView::ColumnView
-        } else {
-            false
-        }
-    } else {
-        false
-    };
-
-    if app.filter.search.is_active && app.mode != AppMode::Search {
-        let search_text = format!("/{}", app.filter.search.query());
-        let help_text = "j/k: navigate | ESC: clear";
-
-        let available_width = area.width.saturating_sub(4);
-        let help_len = help_text.len() as u16;
-        let search_len = search_text.len() as u16;
-
-        let padding = if available_width > search_len + help_len + 1 {
-            available_width
-                .saturating_sub(search_len)
-                .saturating_sub(help_len)
-        } else {
-            1
-        };
-
-        let footer_line = Line::from(vec![
-            Span::styled(search_text, Style::default().fg(Color::Yellow)),
-            Span::styled(
-                format!("{:width$}", "", width = padding as usize),
-                label_text(),
-            ),
-            Span::styled(help_text, label_text()),
-        ]);
-
-        let help = Paragraph::new(footer_line).block(Block::default().borders(Borders::ALL));
-        frame.render_widget(help, area);
-        return;
-    }
-
-    if app.mode == AppMode::Search {
-        let search_text = format!("/{}", app.filter.search.query());
-        let help_text = "ESC: clear | Enter: apply";
-
-        let available_width = area.width.saturating_sub(4);
-        let help_len = help_text.len() as u16;
-        let search_len = search_text.len() as u16;
-
-        let padding = if available_width > search_len + help_len + 1 {
-            available_width
-                .saturating_sub(search_len)
-                .saturating_sub(help_len)
-        } else {
-            1
-        };
-
-        let footer_line = Line::from(vec![
-            Span::styled(search_text, Style::default().fg(Color::White)),
-            Span::styled(
-                format!("{:width$}", "", width = padding as usize),
-                label_text(),
-            ),
-            Span::styled(help_text, label_text()),
-        ]);
-
-        let help = Paragraph::new(footer_line).block(Block::default().borders(Borders::ALL));
-        frame.render_widget(help, area);
-        return;
-    }
-
-    use crate::keybindings::KeybindingRegistry;
-
-    let selection_prefix = if app.multi_select.selection_mode_active {
-        format!(
-            "-- SELECT ({}) -- | ",
-            app.multi_select.selected_cards.len()
-        )
-    } else if !app.multi_select.selected_cards.is_empty() {
-        format!("({} selected) | ", app.multi_select.selected_cards.len())
-    } else {
-        String::new()
-    };
-
-    let help_text: String = if let AppMode::SprintDetail = app.mode {
-        let component = match app.sprint_view.panel {
-            crate::app::SprintTaskPanel::Uncompleted => &app.sprint_view.uncompleted_component,
-            crate::app::SprintTaskPanel::Completed => &app.sprint_view.completed_component,
-        };
-        let provider = KeybindingRegistry::get_provider(app);
-        let context = provider.get_context();
-        let keybindings = context
-            .bindings
-            .iter()
-            .map(|b| format!("{}: {}", b.key, b.short_description))
-            .collect::<Vec<_>>()
-            .join(" | ");
-        let component_help = component.help_text();
-        format!("{}{} | {}", selection_prefix, keybindings, component_help)
-    } else {
-        let provider = KeybindingRegistry::get_provider(app);
-        let context = provider.get_context();
-        let keybindings = context
-            .bindings
-            .iter()
-            .map(|b| format!("{}: {}", b.key, b.short_description))
-            .collect::<Vec<_>>()
-            .join(" | ");
-        format!("{}{}", selection_prefix, keybindings)
-    };
-    let help = Paragraph::new(help_text)
-        .style(label_text())
-        .block(Block::default().borders(Borders::ALL));
-    frame.render_widget(help, area);
-}
-
-fn render_create_board_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Create New Project",
-        "Project Name:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_create_card_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Create New Task",
-        "Task Title:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_create_sprint_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Create New Sprint",
-        "Sprint Name (optional):",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_set_card_points_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Set Points",
-        "Points (1-5 or empty):",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_set_card_priority_popup(app: &App, frame: &mut Frame) {
-    use crate::components::{PriorityDialog, SelectionDialog};
-    let dialog = PriorityDialog;
-    dialog.render(app, frame);
-}
-
-fn render_set_multiple_cards_priority_popup(app: &App, frame: &mut Frame) {
-    use crate::components::{BulkPriorityDialog, SelectionDialog};
-    let dialog = BulkPriorityDialog {
-        count: app.multi_select.selected_cards.len(),
-    };
-    dialog.render(app, frame);
-}
-
-fn render_relationship_boxes(
-    app: &App,
-    frame: &mut Frame,
-    area: Rect,
-    parents: &[Uuid],
-    children: &[Uuid],
-    child_count: usize,
-) {
-    let relationship_chunks = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
-        .split(area);
-
-    let viewport_height =
-        area.height
-            .saturating_sub(RELATIONSHIP_VIEWPORT_BORDER_HEIGHT as u16) as usize;
-
-    // Render Parents section
-    let parents_config = FieldSectionConfig::new("Parents")
-        .with_focus_indicator("Parents [4]")
-        .focused(app.focus.card_focus == CardFocus::Parents);
-    let parents_lines = render_relationship_section(
-        parents,
-        &app.ctx.cards,
-        "Parents",
-        app.focus.card_focus == CardFocus::Parents,
-        &app.relationship.parents_list,
-        viewport_height,
-    );
-    let parents_widget = Paragraph::new(parents_lines).block(parents_config.block());
-    frame.render_widget(parents_widget, relationship_chunks[0]);
-
-    // Render Children section
-    let children_title = format!("Children ({})", child_count);
-    let children_title_focused = format!("Children ({}) [5]", child_count);
-    let children_config = FieldSectionConfig::new(&children_title)
-        .with_focus_indicator(&children_title_focused)
-        .focused(app.focus.card_focus == CardFocus::Children);
-    let children_lines = render_relationship_section(
-        children,
-        &app.ctx.cards,
-        "Children",
-        app.focus.card_focus == CardFocus::Children,
-        &app.relationship.children_list,
-        viewport_height,
-    );
-    let children_widget = Paragraph::new(children_lines).block(children_config.block());
-    frame.render_widget(children_widget, relationship_chunks[1]);
-}
-
-fn render_card_detail_view(app: &App, frame: &mut Frame, area: Rect) {
-    if let Some(card_idx) = app.selection.active_card_index {
-        if let Some(card) = app.ctx.cards.get(card_idx) {
-            if let Some(board_idx) = app.selection.active_board_index {
-                if let Some(board) = app.ctx.boards.get(board_idx) {
-                    let has_sprint_logs = card.sprint_logs.len() > 1;
-                    let card_id = card.id;
-
-                    // Get parent and child information
-                    let parents = app.ctx.graph.cards.parents(card_id);
-                    let children = app.ctx.graph.cards.children(card_id);
-                    let child_count = children.len();
-
-                    let constraints = vec![
-                        Constraint::Length(5),                       // Title
-                        Constraint::Length(6),                       // Metadata
-                        Constraint::Min(5),                          // Description
-                        Constraint::Length(RELATIONSHIP_BOX_HEIGHT), // Relationships
-                    ];
-
-                    let chunks = Layout::default()
-                        .direction(Direction::Vertical)
-                        .constraints(constraints)
-                        .split(area);
-
-                    // Render title section
-                    let title_config = FieldSectionConfig::new("Task Title")
-                        .with_focus_indicator("Task Title [1]")
-                        .focused(app.focus.card_focus == CardFocus::Title);
-                    let title = Paragraph::new(build_title_lines(card))
-                        .style(bold_highlight())
-                        .block(title_config.block());
-                    frame.render_widget(title, chunks[0]);
-
-                    if has_sprint_logs {
-                        let meta_chunks = Layout::default()
-                            .direction(Direction::Horizontal)
-                            .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
-                            .split(chunks[1]);
-
-                        // Render metadata
-                        let meta_config = FieldSectionConfig::new("Metadata")
-                            .with_focus_indicator("Metadata [2]")
-                            .focused(app.focus.card_focus == CardFocus::Metadata);
-                        let meta_lines =
-                            build_metadata_lines(card, board, &app.ctx.sprints, &app.app_config);
-                        let meta = Paragraph::new(meta_lines).block(meta_config.block());
-                        frame.render_widget(meta, meta_chunks[0]);
-
-                        // Render sprint logs
-                        let sprint_logs_config = FieldSectionConfig::new("Sprint History");
-                        let sprint_log_lines = build_sprint_logs_lines(card);
-                        let sprint_logs =
-                            Paragraph::new(sprint_log_lines).block(sprint_logs_config.block());
-                        frame.render_widget(sprint_logs, meta_chunks[1]);
-
-                        // Render description
-                        let desc_config = FieldSectionConfig::new("Description")
-                            .with_focus_indicator("Description [3]")
-                            .focused(app.focus.card_focus == CardFocus::Description);
-                        let desc_lines = build_description_lines(card);
-                        let desc = Paragraph::new(desc_lines).block(desc_config.block());
-                        frame.render_widget(desc, chunks[2]);
-
-                        // Render relationship boxes
-                        render_relationship_boxes(
-                            app,
-                            frame,
-                            chunks[3],
-                            &parents,
-                            &children,
-                            child_count,
-                        );
-                    } else {
-                        // Render metadata section
-                        let meta_config = FieldSectionConfig::new("Metadata")
-                            .with_focus_indicator("Metadata [2]")
-                            .focused(app.focus.card_focus == CardFocus::Metadata);
-                        let meta_lines =
-                            build_metadata_lines(card, board, &app.ctx.sprints, &app.app_config);
-                        let meta = Paragraph::new(meta_lines).block(meta_config.block());
-                        frame.render_widget(meta, chunks[1]);
-
-                        // Render description section
-                        let desc_config = FieldSectionConfig::new("Description")
-                            .with_focus_indicator("Description [3]")
-                            .focused(app.focus.card_focus == CardFocus::Description);
-                        let desc_lines = build_description_lines(card);
-                        let desc = Paragraph::new(desc_lines).block(desc_config.block());
-                        frame.render_widget(desc, chunks[2]);
-
-                        // Render relationship boxes
-                        render_relationship_boxes(
-                            app,
-                            frame,
-                            chunks[3],
-                            &parents,
-                            &children,
-                            child_count,
-                        );
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn render_rename_board_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Rename Project",
-        "New Project Name:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_export_board_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Export Project",
-        "Filename:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_export_all_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Export All Projects",
-        "Filename:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_import_board_popup(app: &App, frame: &mut Frame) {
-    let inner = render_popup_with_block(frame, "Import Projects", 60, 50);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints([Constraint::Length(1), Constraint::Min(0)])
-        .split(inner);
-
-    let label = Paragraph::new("Select a JSON file to import:").style(highlight_text());
-    frame.render_widget(label, chunks[0]);
-
-    if app.dialog_input.import_files.is_empty() {
-        let empty_msg =
-            Paragraph::new("No JSON files found in current directory").style(label_text());
-        frame.render_widget(empty_msg, chunks[1]);
-    } else {
-        let mut lines = vec![];
-        for (idx, filename) in app.dialog_input.import_files.iter().enumerate() {
-            let config = ListItemConfig::new()
-                .selected(app.dialog_input.import_selection.get() == Some(idx))
-                .focused(true);
-            lines.push(styled_list_item(filename, &config));
-        }
-        let list = Paragraph::new(lines);
-        frame.render_widget(list, chunks[1]);
-    }
-}
-
-fn render_board_detail_view(app: &App, frame: &mut Frame, area: Rect) {
-    if let Some(board_idx) = app.selection.board.get() {
-        if let Some(board) = app.ctx.boards.get(board_idx) {
-            let chunks = Layout::default()
-                .direction(Direction::Vertical)
-                .constraints([
-                    Constraint::Length(5),
-                    Constraint::Length(8),
-                    Constraint::Percentage(30),
-                    Constraint::Percentage(35),
-                    Constraint::Percentage(35),
-                ])
-                .split(area);
-
-            let name_config = FieldSectionConfig::new("Project Name")
-                .with_focus_indicator("Project Name [1]")
-                .focused(app.focus.board_focus == BoardFocus::Name);
-            let name = Paragraph::new(board.name.clone())
-                .style(bold_highlight())
-                .block(name_config.block());
-            frame.render_widget(name, chunks[0]);
-
-            let desc_config = FieldSectionConfig::new("Description")
-                .with_focus_indicator("Description [2]")
-                .focused(app.focus.board_focus == BoardFocus::Description);
-            let desc_lines = if let Some(desc_text) = &board.description {
-                crate::markdown_renderer::render_markdown(desc_text)
-            } else {
-                vec![Line::from(Span::styled("No description", label_text()))]
-            };
-            let desc = Paragraph::new(desc_lines).block(desc_config.block());
-            frame.render_widget(desc, chunks[1]);
-
-            let settings_config = FieldSectionConfig::new("Settings")
-                .with_focus_indicator("Settings [3]")
-                .focused(app.focus.board_focus == BoardFocus::Settings);
-
-            let mut settings_lines = vec![
-                if let Some(prefix) = &board.sprint_prefix {
-                    metadata_line_styled("Sprint Prefix", prefix, active_item())
-                } else {
-                    Line::from(vec![
-                        Span::styled("Sprint Prefix: ", label_text()),
-                        Span::styled(
-                            app.app_config.effective_default_sprint_prefix().to_string(),
-                            normal_text(),
-                        ),
-                        Span::styled(" (default)", label_text()),
-                    ])
-                },
-                if let Some(prefix) = &board.card_prefix {
-                    metadata_line_styled("Card Prefix", prefix, active_item())
-                } else {
-                    Line::from(vec![
-                        Span::styled("Card Prefix: ", label_text()),
-                        Span::styled(
-                            app.app_config.effective_default_card_prefix().to_string(),
-                            normal_text(),
-                        ),
-                        Span::styled(" (default)", label_text()),
-                    ])
-                },
-            ];
-
-            // Show active sprint's card prefix override if it exists
-            if let Some(sprint_prefix) =
-                kanban_domain::get_active_sprint_card_prefix_override(board, &app.ctx.sprints)
-            {
-                settings_lines.push(metadata_line_styled(
-                    "Active Sprint Card Prefix",
-                    sprint_prefix,
-                    Style::default().fg(Color::Cyan),
-                ));
-            }
-
-            settings_lines.push(Line::from(""));
-            settings_lines.push(metadata_line(
-                "Sprint Duration",
-                board
-                    .sprint_duration_days
-                    .map(|d| format!("{} days", d))
-                    .unwrap_or_else(|| "(not set)".to_string()),
-            ));
-
-            let available_names: Vec<&str> = board
-                .sprint_names
-                .iter()
-                .skip(board.sprint_name_used_count)
-                .map(|s| s.as_str())
-                .collect();
-
-            if !available_names.is_empty() {
-                settings_lines.push(metadata_line("Sprint Names", available_names.join(", ")));
-            }
-
-            let settings = Paragraph::new(settings_lines).block(settings_config.block());
-            frame.render_widget(settings, chunks[2]);
-
-            let sprints_config = FieldSectionConfig::new("Sprints")
-                .with_focus_indicator("Sprints [4]")
-                .focused(app.focus.board_focus == BoardFocus::Sprints);
-
-            let board_sprints: Vec<&Sprint> = app
-                .ctx
-                .sprints
-                .iter()
-                .filter(|s| s.board_id == board.id)
-                .collect();
-
-            let mut sprint_lines = vec![];
-
-            if board_sprints.is_empty() {
-                sprint_lines.push(Line::from(Span::styled(
-                    "  No sprints yet. Press 'n' to create one!",
-                    label_text(),
-                )));
-            } else {
-                for (sprint_idx, sprint) in board_sprints.iter().enumerate() {
-                    let is_selected = app.selection.sprint.get() == Some(sprint_idx);
-                    let is_focused = app.focus.board_focus == BoardFocus::Sprints;
-
-                    let status_symbol = match sprint.status {
-                        SprintStatus::Planning => "○",
-                        SprintStatus::Active => "●",
-                        SprintStatus::Completed => "✓",
-                        SprintStatus::Cancelled => "✗",
-                    };
-
-                    let sprint_name = sprint.formatted_name(board, "sprint");
-
-                    let card_count = app
-                        .ctx
-                        .cards
-                        .iter()
-                        .filter(|c| c.sprint_id == Some(sprint.id))
-                        .count();
-
-                    let is_active_sprint = board.active_sprint_id == Some(sprint.id);
-                    let is_ended = sprint.is_ended();
-
-                    let mut base_style = normal_text();
-                    if is_selected && is_focused {
-                        base_style = base_style.bg(SELECTED_BG);
-                    }
-
-                    let mut spans = vec![
-                        Span::styled(
-                            format!("{} ", status_symbol),
-                            sprint_status_style(sprint.status),
-                        ),
-                        Span::styled(sprint_name, base_style),
-                        Span::styled(format!(" ({})", card_count), label_text()),
-                    ];
-
-                    if is_active_sprint {
-                        let mut active_style = active_item();
-                        if is_selected && is_focused {
-                            active_style = active_style.bg(SELECTED_BG);
-                        }
-                        spans.push(Span::styled(" Active", active_style));
-                    }
-
-                    if is_ended {
-                        let mut ended_style =
-                            Style::default().fg(Color::Red).add_modifier(Modifier::BOLD);
-                        if is_selected && is_focused {
-                            ended_style = ended_style.bg(SELECTED_BG);
-                        }
-                        spans.push(Span::styled(" Ended", ended_style));
-                    }
-
-                    sprint_lines.push(Line::from(spans));
-                }
-            }
-
-            let sprints = Paragraph::new(sprint_lines).block(sprints_config.block());
-            frame.render_widget(sprints, chunks[3]);
-
-            let columns_config = FieldSectionConfig::new("Columns")
-                .with_focus_indicator("Columns [5]")
-                .focused(app.focus.board_focus == BoardFocus::Columns);
-
-            let mut board_columns: Vec<_> = app
-                .ctx
-                .columns
-                .iter()
-                .filter(|col| col.board_id == board.id)
-                .collect();
-            board_columns.sort_by_key(|col| col.position);
-
-            let mut column_lines = vec![];
-
-            if board_columns.is_empty() {
-                column_lines.push(Line::from(Span::styled(
-                    "  No columns yet. Press 'n' to create one!",
-                    label_text(),
-                )));
-            } else {
-                for (column_idx, column) in board_columns.iter().enumerate() {
-                    let is_selected = app.dialog_input.column_selection.get() == Some(column_idx);
-                    let is_focused = app.focus.board_focus == BoardFocus::Columns;
-
-                    let card_count = app
-                        .ctx
-                        .cards
-                        .iter()
-                        .filter(|c| c.column_id == column.id)
-                        .count();
-
-                    let mut base_style = normal_text();
-                    if is_selected && is_focused {
-                        base_style = base_style.bg(SELECTED_BG);
-                    }
-
-                    let spans = vec![
-                        Span::styled(format!("{}. ", column.position + 1), label_text()),
-                        Span::styled(&column.name, base_style),
-                        Span::styled(format!(" ({})", card_count), label_text()),
-                    ];
-
-                    column_lines.push(Line::from(spans));
-                }
-            }
-
-            let columns = Paragraph::new(column_lines).block(columns_config.block());
-            frame.render_widget(columns, chunks[4]);
-        }
-    }
-}
-
-fn render_set_branch_prefix_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Set Branch Prefix",
-        "Branch Prefix:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_set_sprint_prefix_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Set Sprint Prefix",
-        "Sprint Prefix:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_set_sprint_card_prefix_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Set Card Prefix Override",
-        "Card Prefix:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_order_cards_popup(app: &App, frame: &mut Frame) {
-    use crate::components::{SelectionDialog, SortFieldDialog};
-    let dialog = SortFieldDialog;
-    dialog.render(app, frame);
-}
-
-fn render_assign_sprint_popup(app: &App, frame: &mut Frame) {
-    use crate::components::{SelectionDialog, SprintAssignDialog};
-    let dialog = SprintAssignDialog;
-    dialog.render(app, frame);
-}
-
-fn render_assign_multiple_cards_popup(app: &App, frame: &mut Frame) {
-    let area = centered_rect(60, 50, frame.area());
-
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .title(format!(
-            "Assign {} Cards to Sprint",
-            app.multi_select.selected_cards.len()
-        ))
-        .borders(Borders::ALL)
-        .style(Style::default().bg(Color::Black));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints([Constraint::Length(1), Constraint::Min(0)])
-        .split(inner);
-
-    let label = Paragraph::new("Select sprint:").style(Style::default().fg(Color::Yellow));
-    frame.render_widget(label, chunks[0]);
-
-    let mut lines = vec![];
-
-    if let Some(board_idx) = app.selection.active_board_index {
-        if let Some(board) = app.ctx.boards.get(board_idx) {
-            let board_sprints = Sprint::assignable(&app.ctx.sprints, board.id);
-
-            for (idx, sprint_option) in std::iter::once(None)
-                .chain(board_sprints.iter().map(|s| Some(*s)))
-                .enumerate()
-            {
-                let is_selected = app.dialog_input.sprint_assign_selection.get() == Some(idx);
-
-                let style = if is_selected {
-                    Style::default().fg(Color::White).bg(Color::Blue)
-                } else {
-                    Style::default().fg(Color::White)
-                };
-
-                let prefix = if is_selected { "> " } else { "  " };
-
-                let sprint_name = if let Some(sprint) = sprint_option {
-                    sprint.formatted_name(board, "sprint")
-                } else {
-                    "(None)".to_string()
-                };
-
-                lines.push(Line::from(Span::styled(
-                    format!("{}{}", prefix, sprint_name),
-                    style,
-                )));
-            }
-        }
-    }
-
-    let list = Paragraph::new(lines);
-    frame.render_widget(list, chunks[1]);
-}
-
-fn render_create_column_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Create New Column",
-        "Column Name:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_rename_column_popup(app: &App, frame: &mut Frame) {
-    render_input_popup(
-        frame,
-        "Rename Column",
-        "New Column Name:",
-        app.input.as_str(),
-        app.input.cursor_byte_offset(),
-    );
-}
-
-fn render_delete_column_confirm_popup(_app: &App, frame: &mut Frame) {
-    let area = centered_rect(60, 30, frame.area());
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .title("Delete Column")
-        .borders(Borders::ALL)
-        .style(Style::default().bg(Color::Black));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints([Constraint::Length(2), Constraint::Min(0)])
-        .split(inner);
-
-    let message = Paragraph::new("Are you sure you want to delete this column?\nAll cards will be moved to the first column.")
-        .style(Style::default().fg(Color::Yellow));
-    frame.render_widget(message, chunks[0]);
-
-    let confirm_text =
-        Paragraph::new("Press ENTER/y to delete, n/ESC to cancel").style(label_text());
-    frame.render_widget(confirm_text, chunks[1]);
-}
-
-fn render_select_task_list_view_popup(app: &App, frame: &mut Frame) {
-    use kanban_domain::TaskListView;
-
-    let views = [
-        TaskListView::Flat,
-        TaskListView::GroupedByColumn,
-        TaskListView::ColumnView,
-    ];
-
-    let selected = app.dialog_input.task_list_view_selection.get();
-
-    let current_view = app
-        .selection
-        .active_board_index
-        .and_then(|idx| app.ctx.boards.get(idx))
-        .map(|board| board.task_list_view);
-
-    let items: Vec<ListItem> = views
-        .iter()
-        .enumerate()
-        .map(|(idx, view)| {
-            let style = if Some(idx) == selected {
-                bold_highlight()
-            } else {
-                normal_text()
-            };
-            let is_current = current_view == Some(*view);
-            let view_name = match view {
-                TaskListView::Flat => {
-                    if is_current {
-                        "Flat (current)"
-                    } else {
-                        "Flat"
-                    }
-                }
-                TaskListView::GroupedByColumn => {
-                    if is_current {
-                        "Grouped by Column (current)"
-                    } else {
-                        "Grouped by Column"
-                    }
-                }
-                TaskListView::ColumnView => {
-                    if is_current {
-                        "Column View (kanban board) (current)"
-                    } else {
-                        "Column View (kanban board)"
-                    }
-                }
-            };
-            ListItem::new(view_name).style(style)
-        })
-        .collect();
-
-    render_selection_popup_with_list_items(frame, "Select Task List View", items, 50, 40);
-}
-
-fn render_filter_options_popup(app: &App, frame: &mut Frame) {
-    let area = centered_rect(70, 75, frame.area());
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .title("Filter Options")
-        .borders(Borders::ALL)
-        .border_style(focused_border());
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints([
-            Constraint::Min(8),
-            Constraint::Length(3),
-            Constraint::Length(3),
-        ])
-        .split(inner);
-
-    if let Some(ref dialog_state) = app.filter.dialog_state {
-        let section_index = dialog_state.section_index;
-
-        let mut sprint_lines = vec![Line::from(Span::styled(
-            "Sprints",
-            if section_index == 0 {
-                bold_highlight()
-            } else {
-                normal_text()
-            },
-        ))];
-
-        let unassigned_cursor = if section_index == 0 && dialog_state.item_selection == 0 {
-            "> "
-        } else {
-            "  "
-        };
-
-        sprint_lines.push(Line::from(vec![
-            Span::raw(unassigned_cursor),
-            Span::styled(
-                if dialog_state.filters.show_unassigned_sprints {
-                    "[✓]"
-                } else {
-                    "[ ]"
-                },
-                normal_text(),
-            ),
-            Span::raw(" "),
-            Span::styled("Show cards with unassigned sprints", normal_text()),
-        ]));
-
-        sprint_lines.push(Line::from(Span::styled(
-            "─────────────────────────",
-            label_text(),
-        )));
-
-        if let Some(board_idx) = app.selection.active_board_index {
-            if let Some(board) = app.ctx.boards.get(board_idx) {
-                let board_sprints: Vec<_> = app
-                    .ctx
-                    .sprints
-                    .iter()
-                    .filter(|s| s.board_id == board.id)
-                    .collect();
-
-                if board_sprints.is_empty() {
-                    sprint_lines.push(Line::from(Span::styled(
-                        "  (no sprints available)",
-                        label_text(),
-                    )));
-                } else {
-                    for (idx, sprint) in board_sprints.iter().enumerate() {
-                        let is_selected = dialog_state
-                            .filters
-                            .selected_sprint_ids
-                            .contains(&sprint.id);
-                        let cursor = if section_index == 0 && dialog_state.item_selection == idx + 1
-                        {
-                            "> "
-                        } else {
-                            "  "
-                        };
-
-                        sprint_lines.push(Line::from(vec![
-                            Span::raw(cursor),
-                            Span::styled(if is_selected { "[✓]" } else { "[ ]" }, normal_text()),
-                            Span::raw(" "),
-                            Span::styled(sprint.formatted_name(board, "sprint"), normal_text()),
-                        ]));
-                    }
-                }
-            }
-        }
-
-        let section1 = Paragraph::new(sprint_lines).block(
-            Block::default()
-                .borders(Borders::ALL)
-                .border_style(if section_index == 0 {
-                    focused_border()
-                } else {
-                    Style::default()
-                }),
-        );
-        frame.render_widget(section1, chunks[0]);
-
-        let date_lines = vec![
-            Line::from(Span::styled(
-                "Date Range (Future)",
-                if section_index == 1 {
-                    bold_highlight()
-                } else {
-                    label_text()
-                },
-            )),
-            Line::from(Span::styled(
-                "  Filter by last updated or created date",
-                label_text(),
-            )),
-        ];
-
-        let section2 =
-            Paragraph::new(date_lines).block(Block::default().borders(Borders::ALL).border_style(
-                if section_index == 1 {
-                    focused_border()
-                } else {
-                    Style::default()
-                },
-            ));
-        frame.render_widget(section2, chunks[1]);
-
-        let tag_lines = vec![
-            Line::from(Span::styled(
-                "Tags (Future)",
-                if section_index == 2 {
-                    bold_highlight()
-                } else {
-                    label_text()
-                },
-            )),
-            Line::from(Span::styled("  Filter cards by tags", label_text())),
-        ];
-
-        let section3 =
-            Paragraph::new(tag_lines).block(Block::default().borders(Borders::ALL).border_style(
-                if section_index == 2 {
-                    focused_border()
-                } else {
-                    Style::default()
-                },
-            ));
-        frame.render_widget(section3, chunks[2]);
-    }
-}
-
-pub fn help_popup_viewport_height(frame_area: Rect) -> usize {
-    let popup = centered_rect(80, 80, frame_area);
-    let block = Block::default().borders(Borders::ALL);
-    let inner = block.inner(popup);
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .horizontal_margin(2)
-        .constraints([
-            Constraint::Length(2),
-            Constraint::Min(0),
-            Constraint::Length(2),
-        ])
-        .split(inner);
-    chunks[1].height as usize
-}
-
-fn render_help_popup(app: &App, frame: &mut Frame) {
-    use crate::components::ListItemConfig;
-    use crate::keybindings::KeybindingRegistry;
-
-    let area = centered_rect(80, 80, frame.area());
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .title("Help - Keybindings for Current Context")
-        .borders(Borders::ALL)
-        .border_style(focused_border());
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    // Layout: fixed header (context name + blank), scrollable bindings, fixed footer
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .horizontal_margin(2)
-        .constraints([
-            Constraint::Length(2),
-            Constraint::Min(0),
-            Constraint::Length(2),
-        ])
-        .split(inner);
-
-    // Get keybindings for the underlying mode
-    let provider = KeybindingRegistry::get_provider(app);
-    let context = provider.get_context();
-
-    // Fixed header
-    frame.render_widget(
-        Paragraph::new(vec![
-            Line::from(Span::styled(
-                context.name.clone(),
-                Style::default()
-                    .fg(Color::Cyan)
-                    .add_modifier(Modifier::BOLD),
-            )),
-            Line::from(""),
-        ]),
-        chunks[0],
-    );
-
-    // Scrollable bindings body
-    let raw_height = help_popup_viewport_height(frame.area());
-
-    let selected_idx = app.ui_state.help_list.get_selected_index();
-
-    let adjusted_height = app
-        .ui_state
-        .help_list
-        .get_adjusted_viewport_height(raw_height);
-    let page_info = app.ui_state.help_list.get_render_info(adjusted_height);
-
-    let mut rendered_lines: Vec<Line> = crate::scroll_indicators::render_above_indicator(
-        page_info.show_above_indicator,
-        page_info.items_above,
-        "item",
-    )
-    .into_iter()
-    .collect();
-
-    let visible_lines: Vec<Line> = page_info
-        .visible_indices
-        .iter()
-        .filter_map(|&i| {
-            let binding = context.bindings.get(i)?;
-            let is_selected = selected_idx == Some(i);
-            let config = ListItemConfig::new().selected(is_selected).focused(true);
-            let prefix = config.item_prefix();
-            let style = config.item_style();
-            Some(Line::from(vec![
-                Span::styled(prefix.to_string(), style),
-                Span::styled(binding.key.to_string(), Style::default().fg(Color::Yellow)),
-                Span::raw(" "),
-                Span::styled(binding.description.clone(), style),
-            ]))
-        })
-        .collect();
-    rendered_lines.extend(visible_lines);
-    rendered_lines.extend(crate::scroll_indicators::render_below_indicator(
-        page_info.show_below_indicator,
-        page_info.items_below,
-        "item",
-    ));
-
-    frame.render_widget(Paragraph::new(rendered_lines), chunks[1]);
-
-    // Fixed footer — always visible regardless of scroll position
-    let footer = Paragraph::new(vec![
-        Line::from(""),
-        Line::from(Span::styled(
-            "j/k or ↑↓: navigate | Enter: activate | ESC or ?: close",
-            Style::default()
-                .fg(Color::DarkGray)
-                .add_modifier(Modifier::ITALIC),
-        )),
-    ]);
-    frame.render_widget(footer, chunks[2]);
-}
-
-fn render_conflict_resolution_popup(_app: &App, frame: &mut Frame) {
-    let area = centered_rect(70, 40, frame.area());
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .title("File Conflict Detected")
-        .borders(Borders::ALL)
-        .style(Style::default().bg(Color::Black));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints([
-            Constraint::Length(3),
-            Constraint::Length(4),
-            Constraint::Min(0),
-        ])
-        .split(inner);
-
-    let message = Paragraph::new(
-        "The file was modified by another instance.\nChoose how to resolve this conflict:",
-    )
-    .style(Style::default().fg(Color::Yellow));
-    frame.render_widget(message, chunks[0]);
-
-    let options = vec![
-        Line::from(Span::styled(
-            "(O)verwrite",
-            Style::default().fg(Color::Cyan),
-        )),
-        Line::from(Span::styled(
-            "  Keep your changes and overwrite the file",
-            label_text(),
-        )),
-        Line::from(""),
-        Line::from(Span::styled(
-            "(T)ake theirs",
-            Style::default().fg(Color::Cyan),
-        )),
-        Line::from(Span::styled(
-            "  Discard your changes and reload the file",
-            label_text(),
-        )),
-    ];
-    let options_para = Paragraph::new(options);
-    frame.render_widget(options_para, chunks[1]);
-
-    let instructions =
-        Paragraph::new("Press O or T to choose, ESC to retry later").style(label_text());
-    frame.render_widget(instructions, chunks[2]);
-}
-
-fn render_external_change_detected_popup(_app: &App, frame: &mut Frame) {
-    let area = centered_rect(70, 40, frame.area());
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .title("External File Change Detected")
-        .borders(Borders::ALL)
-        .style(Style::default().bg(Color::Black));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(2)
-        .constraints([
-            Constraint::Length(3),
-            Constraint::Length(4),
-            Constraint::Min(0),
-        ])
-        .split(inner);
-
-    let message = Paragraph::new(
-        "The file was modified by another instance.\nYou have unsaved changes. Choose an action:",
-    )
-    .style(Style::default().fg(Color::Yellow));
-    frame.render_widget(message, chunks[0]);
-
-    let options = vec![
-        Line::from(Span::styled("(R)eload", Style::default().fg(Color::Cyan))),
-        Line::from(Span::styled(
-            "  Discard your changes and reload the file",
-            label_text(),
-        )),
-        Line::from(""),
-        Line::from(Span::styled("(K)eep", Style::default().fg(Color::Cyan))),
-        Line::from(Span::styled(
-            "  Continue with your changes (save will overwrite)",
-            label_text(),
-        )),
-    ];
-    let options_para = Paragraph::new(options);
-    frame.render_widget(options_para, chunks[1]);
-
-    let instructions =
-        Paragraph::new("Press R or K to choose, ESC to continue").style(label_text());
-    frame.render_widget(instructions, chunks[2]);
-}
-
-fn render_manage_parents_popup(app: &App, frame: &mut Frame) {
-    render_relationship_popup(app, frame, "Set Parents", true);
-}
-
-fn render_manage_children_popup(app: &App, frame: &mut Frame) {
-    render_relationship_popup(app, frame, "Set Children", false);
-}
-
-fn render_carry_over_sprint_popup(app: &App, frame: &mut Frame) {
-    use crate::components::selection_dialog::CarryOverSprintDialog;
-    use crate::components::SelectionDialog;
-    let card_count = app
-        .dialog_input
-        .carry_over_source_sprint_id
-        .map(|id| {
-            use kanban_domain::query::sprint::get_sprint_uncompleted_cards;
-            get_sprint_uncompleted_cards(id, &app.ctx.cards).len()
-        })
-        .unwrap_or(0);
-    CarryOverSprintDialog { card_count }.render(app, frame);
-}
-
-fn render_relationship_popup(app: &App, frame: &mut Frame, title: &str, _is_parent_mode: bool) {
-    use crate::components::centered_rect;
-    use ratatui::widgets::Clear;
-
-    let area = centered_rect(60, 70, frame.area());
-    frame.render_widget(Clear, area);
-
-    let block = Block::default()
-        .title(title)
-        .borders(Borders::ALL)
-        .style(Style::default().bg(Color::Black));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .margin(1)
-        .constraints([
-            Constraint::Length(3), // Search box
-            Constraint::Min(0),    // Card list
-            Constraint::Length(1), // Instructions
-        ])
-        .split(inner);
-
-    // Search box
-    let search_border_style = if app.relationship.search_active {
-        Style::default().fg(Color::Yellow)
-    } else {
-        Style::default().fg(Color::Gray)
-    };
-    let search_block = Block::default()
-        .title("Search")
-        .borders(Borders::ALL)
-        .border_style(search_border_style);
-
-    // Build search text with greyed "/" prefix when not in search mode
-    let search_text: Line = if app.relationship.search_active {
-        // In search mode - show cursor indicator
-        Line::from(vec![
-            Span::styled(&app.relationship.search, Style::default().fg(Color::White)),
-            Span::styled("_", Style::default().fg(Color::Yellow)),
-        ])
-    } else if app.relationship.search.is_empty() {
-        // Not in search mode, empty search - show greyed "/" hint
-        Line::from(Span::styled(
-            "/ to search",
-            Style::default().fg(Color::DarkGray),
-        ))
-    } else {
-        // Not in search mode, has search text
-        Line::from(Span::styled(
-            &app.relationship.search,
-            Style::default().fg(Color::White),
-        ))
-    };
-
-    let search = Paragraph::new(search_text).block(search_block);
-    frame.render_widget(search, chunks[0]);
-
-    // Filter cards by search
-    let filtered_cards: Vec<_> = if app.relationship.search.is_empty() {
-        app.relationship.card_ids.clone()
-    } else {
-        let search_lower = app.relationship.search.to_lowercase();
-        app.relationship
-            .card_ids
-            .iter()
-            .filter(|card_id| {
-                app.ctx
-                    .cards
-                    .iter()
-                    .find(|c| c.id == **card_id)
-                    .map(|c| c.title.to_lowercase().contains(&search_lower))
-                    .unwrap_or(false)
-            })
-            .copied()
-            .collect()
-    };
-
-    // Card list
-    let mut lines = vec![];
-    for (idx, card_id) in filtered_cards.iter().enumerate() {
-        if let Some(card) = app.ctx.cards.iter().find(|c| c.id == *card_id) {
-            let is_selected = app.relationship.selection.get() == Some(idx);
-            let is_checked = app.relationship.selected.contains(card_id);
-
-            let checkbox = if is_checked { "[✓]" } else { "[ ]" };
-
-            let style = if is_selected {
-                Style::default().fg(Color::White).bg(Color::Blue)
-            } else if is_checked {
-                Style::default().fg(Color::Green)
-            } else {
-                Style::default().fg(Color::White)
-            };
-
-            lines.push(Line::from(Span::styled(
-                format!("{} {}", checkbox, card.title),
-                style,
-            )));
-        }
-    }
-
-    if lines.is_empty() {
-        lines.push(Line::from(Span::styled(
-            "No eligible cards found",
-            Style::default().fg(Color::DarkGray),
-        )));
-    }
-
-    let list = Paragraph::new(lines);
-    frame.render_widget(list, chunks[1]);
-
-    // Instructions
-    let instructions_text = if app.relationship.search_active {
-        "Type to search | Enter/Esc: exit search"
-    } else {
-        "j/k: navigate | Space: toggle | /: search | Esc: close"
-    };
-    let instructions =
-        Paragraph::new(instructions_text).style(Style::default().fg(Color::DarkGray));
-    frame.render_widget(instructions, chunks[2]);
-}
